backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-socketio==5.11.3
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
psycopg2-binary==2.9.9
pymongo==4.8.0
redis==5.0.7
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
email-validator==2.2.0
itsdangerous==2.2.0
python-multipart==0.0.9
openai==1.42.0
httpx==0.27.0

# Google Gemini SDK
google-generativeai==0.7.2

